add ref check

+++ handlers/users/command_start.py
@@ -1,10 +1,10 @@
-from asyncio.windows_events import NULL
 from datetime import timedelta
 from aiogram import types
 from aiogram.dispatcher.filters.builtin import CommandStart
 from aiogram.types import CallbackQuery
 from loader import dp, bot
 from keyboards.default import menu
+from keyboards.default import y_n
 from filters.emoji import *
 from database.classes import *
 import random
@@ -15,22 +15,41 @@ path = database.path
 
 @dp.message_handler(CommandStart())
 async def bot_start(message: types.Message):
-    
-
-
-    
+        
     user_id = message.from_user.id
     name = message.from_user.first_name
     user_name = message.from_user.username
 
     database.write_in_db(path,'users',[user_id, name, user_name])
+    sub_format = database.find_value(path, "users", message.from_user.id,'sub_format')
     referrer_code = database.find_value(path, 'users', message.from_user.id, 'referrer_code')
     rand = random.randint(1000000000000000,9999999999999999)
     if user_id != 'Empty' and referrer_code == 'Empty':
         database.change_value(path,'users', message.from_user.id, 'referrer_code', rand)
-    await message.delete()
+    if sub_format == 'Free':
+        await message.answer('У вас есть реферальный код?', reply_markup=y_n)
+    else:
+        text = f'<b>Омский Поликек</b> приветсвует тебя, {message.from_user.full_name} 👋'
+        await message.answer(text, reply_markup=menu)
+        await message.delete()
+
+
+@dp.message_handler(text=f'Да')
+async def bot_data_request(message: types.Message):
+    await message.answer("Введите реферальный код")
+
+
+@dp.message_handler(text=f'Нет')
+async def bot_data_request(message: types.Message):
     text = f'<b>Омский Поликек</b> приветсвует тебя, {message.from_user.full_name} 👋'
     await message.answer(text, reply_markup=menu)
+
+@dp.message_handler(text_contains = '1')
+async def bot_data_request(message: types.Message):
+    text = database.check_ref_code(path, "users", message.text, 'referrer_code')
+    await message.answer(text, reply_markup=menu)
+    await message.answer(f'<b>Омский Поликек</b> приветсвует тебя, {message.from_user.full_name} 👋')
+
 
 
 # refferrer_code = database.find_value(path, 'users', 'referrer_code')
